#75 - Add GameAi
