+++ apps/lundin/src/app/modules/virus/virus-game/virus-game.component.ts
@@ -1,4 +1,5 @@
 import { Component, ElementRef, HostBinding, Input } from "@angular/core"
+import { GameAi } from "@lundin/age"
 import { range } from "@lundin/utility"
 import { Virus, VirusAction, VirusConfig } from "@lundin/virus"
 
@@ -10,7 +11,7 @@ import { Virus, VirusAction, VirusConfig } from "@lundin/virus"
 export class VirusGameComponent {
 	@Input() players = [
 		new VirusPlayer("Spiller 1", "red"),
-		new VirusPlayer("Spiller 2", "green"),
+		new VirusPlayer("Spiller 2", "green", new SimpleAi()),
 	]
 	@Input() boardSize = 8
 	@Input() fieldSize = 50
@@ -96,7 +97,10 @@ export class VirusGameComponent {
 	}
 
 	private announceNextTurn() {
-		this.message = this.getPlayer().name + "'s tur"
+		const player = this.getPlayer()
+		this.message = player.name + "'s tur"
+		if (player.ai)
+			setTimeout(() => this.performAction(player.ai.requestActions(this.game)[0]), 100)
 	}
 
 	colorFor(position: { x: number, y: number }) {
@@ -119,5 +123,12 @@ export class VirusPlayer {
 	constructor(
 		public name: string,
 		public color: string,
+		public ai: GameAi<Virus, VirusAction> = null
 	) { }
+}
+
+export class SimpleAi implements GameAi<Virus, VirusAction>{
+	requestActions(game: Virus) {
+		return [new VirusAction(game.state.currentPlayer, 0, 7, 0, 6)]
+	}
 }
