+++ libs/age/src/index.ts
@@ -1,2 +1,6 @@
 export * from "./lib/age"
+export * from "./lib/game-ai"
 export * from "./lib/game-grid"
+export * from "./lib/game-logic"
+export * from "./lib/game-state"
+export * from "./lib/game-validator"
