+++ libs/age/src/lib/game-ai.ts
@@ -1,0 +1,3 @@
+export interface GameAi<Game, GameAction> {
+	requestActions(game: Game): GameAction[]
+}
