+++ libs/age/src/lib/game-logic.ts
@@ -1,0 +1,3 @@
+export interface GameLogic<GameAction> {
+	update(actions: GameAction[]): void
+}
