+++ libs/age/src/lib/game-state.ts
@@ -1,0 +1,4 @@
+export interface GameState {
+	tick: number
+	finishUpdate?(): void
+}
