+++ libs/age/src/lib/game-validator.ts
@@ -1,0 +1,8 @@
+export interface GameValidator<GameAction> {
+	validate(actions: GameAction[]): Validation
+}
+
+export interface Validation {
+	isValid: boolean
+	problems: string[]
+}
